Moved templates to their corresponding model folder
