+++ shopyo/templates/base/nav_base.html
@@ -12,8 +12,10 @@
 
 <nav class="navbar navbar-expand-sm bg-dark navbar-dark">
   <!-- Brand/logo -->
-  <a class="navbar-brand" href="#"> <img src="/static/shopyo.svg" width="35" heigh
-    t="30"> {{OUR_APP_NAME}} | {{active_page.capitalize()}}</a>
+  <a class="navbar-brand" href="#">
+    <img src="/static/shopyo.svg" width="35" height="30">
+    {{OUR_APP_NAME}} | {{active_page.capitalize()}}
+  </a>
 
   <!-- Links -->
   <ul class="navbar-nav">
