+++ shopyo/views/admin/templates/admin/index.html
