+++ shopyo/views/manufacturer/templates/manufac/edit.html
