+++ shopyo/views/manufacturer/templates/manufac/index.html
