+++ shopyo/views/people/templates/people/index.html
