+++ shopyo/views/products/templates/prods/add.html
